undo rename. May revert
