+++ .gitignore
@@ -101,3 +101,7 @@ venv.bak/
 static/
 .idea/*
 */media/wiki/*
+
+# build artifacts
+buttercup
+setup.py
