+++ buttercup/cogs/__init__.py
