+++ buttercup/cogs/heatmap.py
@@ -11,8 +11,8 @@ from discord.ext.commands import Cog
 from discord_slash import SlashContext, cog_ext
 from discord_slash.utils.manage_commands import create_option
 
-from src.bot import ButtercupBot
-from src.cogs.helpers import (
+from buttercup.bot import ButtercupBot
+from buttercup.cogs.helpers import (
     BlossomException,
     BlossomUser,
     extract_utc_offset,
@@ -24,7 +24,7 @@ from src.cogs.helpers import (
     parse_time_constraints,
     utc_offset_to_str,
 )
-from src.strings import translation
+from buttercup.strings import translation
 
 i18n = translation()
