+++ buttercup/cogs/history.py
@@ -15,9 +15,9 @@ from discord_slash import SlashContext, cog_ext
 from discord_slash.model import SlashMessage
 from discord_slash.utils.manage_commands import create_option
 
-from src.bot import ButtercupBot
-from src.cogs import ranks
-from src.cogs.helpers import (
+from buttercup.bot import ButtercupBot
+from buttercup.cogs import ranks
+from buttercup.cogs.helpers import (
     BlossomException,
     BlossomUser,
     InvalidArgumentException,
@@ -38,7 +38,7 @@ from src.cogs.helpers import (
     parse_time_constraints,
     utc_offset_to_str,
 )
-from src.strings import translation
+from buttercup.strings import translation
 
 i18n = translation()
