+++ buttercup/cogs/ping.py
@@ -5,7 +5,7 @@ from discord import Color, Embed
 from discord.ext.commands import Cog
 from discord_slash import SlashContext, cog_ext
 
-from src.bot import ButtercupBot
+from buttercup.bot import ButtercupBot
 
 
 class Ping(Cog):
