+++ buttercup/strings/__init__.py
