+++ pyproject.toml
@@ -1,7 +1,7 @@
 [tool.poetry]
 name = "buttercup"
 version = "0.24.0"
-description = "The Discord Bot!"
+description = "Buttercup: The Discord Bot!"
 authors = ["Grafeas Group Ltd. <devs@grafeas.org>"]
 exclude = ['test', 'test.*', '*.test.*', '*.test']
 
@@ -36,6 +36,7 @@ pytest = "^7.1.2"
 pytest-cov = "^3.0.0"
 coveralls = "^3.2.0"
 flake8-isort = "^4.1.2"
+poetry = "^1.1.14"
 
 [tool.poetry.extras]
 ci = ["pytest", "pytest-cov", "coveralls"]
@@ -43,7 +44,7 @@ ci = ["pytest", "pytest-cov", "coveralls"]
 [tool.isort]
 profile = "black"
 known_first_party = "src"
-known_third_party = ["asyncpraw", "asyncprawcore", "blossom_wrapper", "dateutil", "discord", "discord_slash", "matplotlib", "pandas", "pytest", "pytz", "requests", "seaborn", "toml", "yaml"]
+known_third_party = ["asyncpraw", "asyncprawcore", "blossom_wrapper", "dateutil", "discord", "discord_slash", "matplotlib", "pandas", "poetry", "pytest", "pytz", "requests", "seaborn", "toml", "yaml"]
 
 [build-system]
 requires = ["poetry>=0.12"]
