+++ test/test_helpers.py
@@ -4,7 +4,7 @@ from typing import List, Optional
 import pytz
 from pytest import mark
 
-from src.cogs.helpers import (
+from buttercup.cogs.helpers import (
     BlossomUser,
     escape_formatting,
     extract_sub_name,
